requirements.txt
# CRISP-DM Machine Learning Pipeline with RLT Methodology
# Requirements file for Python dependencies

# Core Data Science Libraries
pandas>=1.5.0
numpy>=1.23.0
scipy>=1.9.0

# Machine Learning
scikit-learn>=1.2.0
xgboost>=1.7.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Document Processing (for reading project docs)
python-docx>=0.8.11
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# Jupyter Notebook (optional)
jupyter>=1.0.0
notebook>=6.5.0
ipykernel>=6.20.0

# Excel Support
openpyxl>=3.0.0
xlrd>=2.0.0

# Utilities
tqdm>=4.64.0
joblib>=1.2.0

# Additional Dependencies
cryptography>=36.0.0  # For secure model saving
Pillow>=9.1.0  # For image processing in visualizations
